requirements.txt
discord.py>=2.0.0
requests>=2.27.1
flask>=3.0.0
discord-py-interactions
discord-py-slash-command
discord.py>=2.0.0
